toolkit/content/tests/chrome/window_popup_preventdefault_chrome.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>

<window title="Popup Prevent Default Tests"
  onload="setTimeout(runTest, 0);"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <title>Popup Prevent Default Tests</title>

<!--
  This tests checks that preventDefault can be called on a popupshowing
  event or popuphiding event to prevent the default behaviour.
  -->

<script>

var gBlockShowing = true;
var gBlockHiding = true;
var gShownNotAllowed = true;
var gHiddenNotAllowed = true;

function runTest()
{
  document.getElementById("menu").open = true;
}

function popupShowing(event)
{
  if (gBlockShowing) {
    event.preventDefault();
    gBlockShowing = false;
    setTimeout(function() {
                 gShownNotAllowed = false;
                 document.getElementById("menu").open = true;
               }, 3000, true);
  }
}

function popupShown()
{
  window.opener.wrappedJSObject.SimpleTest.ok(!gShownNotAllowed, "popupshowing preventDefault");
  document.getElementById("menu").open = false;
}

function popupHiding(event)
{
  if (gBlockHiding) {
    event.preventDefault();
    gBlockHiding = false;
    setTimeout(function() {
                 gHiddenNotAllowed = false;
                 document.getElementById("menu").open = false;
               }, 3000, true);
  }
}

function popupHidden()
{
  window.opener.wrappedJSObject.SimpleTest.ok(!gHiddenNotAllowed, "popuphiding preventDefault");
  window.opener.wrappedJSObject.SimpleTest.finish();
  window.close();
}
</script>

<button id="menu" type="menu" label="Menu">
  <menupopup onpopupshowing="popupShowing(event);"
             onpopupshown="popupShown();"
             onpopuphiding="popupHiding(event);"
             onpopuphidden="popupHidden();">
    <menuitem label="Item"/>
  </menupopup>
</button>


</window>
